+++ packages/typespec-go/CHANGELOG.md
@@ -6,6 +6,10 @@
 
 * Fix polymorphic type result of response generation.
 
+### Bugs Fixed
+
+* Fixed some edge cases where a method parameter's optionality wasn't correctly handled.
+
 ### Other Changes
 
 * Updated the minimum version of `azcore` to `v1.19.0`.